done group

+++ server/src/http/models/myfile.cr
@@ -103,23 +103,76 @@ module DMACServer
         File.write(@full_path, text)
       end
 
-      def self.collect_files(role, project, data_path)
+      def self.collect_files(role, group, project, data_path)
         files = [] of MyFile
+        raise "You do not have permission" unless MyFile.check(role, group, data_path)
         dir = MyFile.new(project, data_path)
         files << dir
         return files if dir.type.to_s == "file"
         Dir.foreach dir.full_path do |filename|
-          if filename.to_s != "." && filename.to_s != ".." && filename.to_s != ".git"
+          if filename.to_s != "." && filename.to_s != ".." && filename.to_s != ".git" && filename.to_s != ".gitignore"
             dp = filename
             dp = data_path + "--" + dp unless data_path == "-root-"
-            if filename.to_s[0] != '.'
-              files << MyFile.new(project, dp)
-            elsif role == "Owner" || role == "Admin"
+            if MyFile.check(role, group, dp)
               files << MyFile.new(project, dp)
             end
           end
         end
         return files
+      end
+
+      def self.check(role, group, data_path)
+        return true if data_path == "-root-"
+        return true if role == "Owner" || role == "Admin"
+        return false if MyFile.is_hidden(data_path)
+        return true if group == ""
+        fg = MyFile.group_own(data_path)
+        return true if fg == ""
+        return fg == group
+      end
+
+      def self.check_full_path(control, project_root, full_path)
+        data_path = "-root-"
+        if project_root != full_path
+          data_path = full_path[project_root.size+1..-1].gsub("/", "--")
+        end
+        return true if data_path == "-root-"
+        return true if control.role.to_s == "Owner" || control.role.to_s == "Admin"
+        return false if MyFile.is_hidden(data_path)
+        return true if control.group_name.to_s == ""
+        fg = MyFile.group_own(data_path)
+        return true if fg == ""
+        return fg == control.group_name.to_s
+      end
+
+      def self.is_hidden(data_path)
+        hidden = false
+        data_path.split("--") do |s|
+          if s[0] == '.'
+            hidden = true
+            break
+          end
+        end
+        return hidden
+      end
+
+      def self.group_own(data_path)
+        group = ""
+        data_path.split("--") do |s|
+          g = ""
+          i = 0
+          s.split("_") do |ss|
+            break if i == 0 && ss != ""
+            g = ss if i == 1
+            i = i + 1
+            break if i >= 3
+          end
+          if g != "" && i >= 3
+            group = g
+            break
+          end
+        end
+        return group
       end
 
       def self.create_folder(project, data_path)
@@ -163,23 +216,28 @@ module DMACServer
         end
       end
 
-      def self.update_folder_file_name(project, data_path, name)
+      def self.update_folder_file_name(project, data_path, name, control)
         full_path = @@root + "/" + project.key.to_s + "/" + data_path.gsub("--", "/")
         raise "Cannot find file" unless File.exists?(full_path)
+        raise "No permission" unless MyFile.check(control.role.to_s, control.group_name.to_s, data_path)
         dirname = File.dirname(full_path)
         new_full_path = dirname.to_s + "/" + name
         File.rename(full_path, new_full_path)
       end
 
-      def self.delete_folder_file(project, data_path)
+      def self.delete_folder_file(project, data_path, control)
+        raise "No permission" unless MyFile.check(control.role.to_s, control.group_name.to_s, data_path)
         full_path = @@root + "/" + project.key.to_s + "/" + data_path.gsub("--", "/")
         MyFile.delete_files(full_path)
       end
 
-      def self.upload_file(project, data_path, file)
+      def self.upload_file(project, data_path, file, control)
         full_path = @@root + "/" + project.key.to_s
         prefix_length = full_path.size + 1
         full_path = full_path + "/" + data_path.gsub("--", "/") if data_path != "-root-"
+
+        raise "Cannot find file" unless File.exists?(full_path)
+        raise "No permission" unless MyFile.check(control.role.to_s, control.group_name.to_s, data_path)
 
         filename = file.filename
         raise "No filename included in upload" if !filename.is_a?(String)
@@ -237,24 +295,29 @@ module DMACServer
         return full_path
       end
 
-      def self.save_text_file(project, data_path, text)
+      def self.save_text_file(project, data_path, text, control)
+        raise "No permission" unless MyFile.check(control.role.to_s, control.group_name.to_s, data_path)
         file = MyFile.new(project, data_path)
         raise "No such file" unless file.fileType == "text"
         file.save_text(text)
       end
 
-      def self.copy_files(source_files, target_file)
+      def self.copy_files(source_files, target_file, source_control, target_control)
+        raise "No permission" unless MyFile.check(target_control.role.to_s, target_control.group_name.to_s, target_file.data_path)
+
         new_folders = {} of String => Bool
         source_files.each do |f|
+          project_root = @@root + "/" + f.project.key.to_s
           if f.type == "file"
-            MyFile.copy_file(f.full_path, target_file.full_path)
+            MyFile.copy_file(f.full_path, target_file.full_path, source_control, project_root)
           else
-            MyFile.copy_folder(f.full_path, target_file.full_path, new_folders)
+            MyFile.copy_folder(f.full_path, target_file.full_path, new_folders, source_control, project_root)
           end
         end
       end
 
-      def self.copy_file(source_full_path, target_full_path)
+      def self.copy_file(source_full_path, target_full_path, control, project_root)
+        return unless MyFile.check_full_path(control, project_root, source_full_path)
         new_full_path = MyFile.make_name(source_full_path, target_full_path)
         File.open(new_full_path, "w") do |tf|
           File.open(source_full_path) do |sf|
@@ -263,7 +326,8 @@ module DMACServer
         end
       end
 
-      def self.copy_folder(source_full_path, target_full_path, new_folders)
+      def self.copy_folder(source_full_path, target_full_path, new_folders, control, project_root)
+        return unless MyFile.check_full_path(control, project_root, source_full_path)
         new_target_full_path = MyFile.make_name(source_full_path, target_full_path)
         Dir.mkdir(new_target_full_path)
         new_folders[new_target_full_path] = true
@@ -271,9 +335,9 @@ module DMACServer
           if filename.to_s != "." && filename.to_s != ".."
             new_source_full_path = source_full_path + "/" + filename
             if File.file? new_source_full_path
-              MyFile.copy_file(new_source_full_path, new_target_full_path)
+              MyFile.copy_file(new_source_full_path, new_target_full_path, control, project_root)
             elsif (File.directory? new_source_full_path) && (!new_folders.has_key?(new_source_full_path))
-              MyFile.copy_folder(new_source_full_path, new_target_full_path, new_folders)
+              MyFile.copy_folder(new_source_full_path, new_target_full_path, new_folders, control, project_root)
             end
           end
         end
